envs: Cast action bounds to float32
After the following warning:

    env_checker.py:473: UserWarning: Your action space has dtype
    float64, we recommend using np.float32 to avoid cast errors.

+++ upkie/envs/wrappers/pendulum.py
@@ -140,7 +140,7 @@ class Pendulum(gym.Wrapper):
             ],
             dtype=float,
         )
-        action_limit = np.array([max_ground_velocity], dtype=float)
+        action_limit = np.array([max_ground_velocity], dtype=np.float32)
 
         # gymnasium.Env: observation_space
         self.observation_space = gym.spaces.Box(
